+++ src/template/ITemplateProps.ts
@@ -7,6 +7,11 @@ export interface ITemplateProps {
 
     outputMapping?: IOutputType;
 
+    /**
+     * Below option has no effect on the output directory path if used without
+     * calling cliRenderer. CliRenderer after setting everything up will set
+     * value from this property to CWD variable in the root scope.
+     */
     defaultOutputDirectoryPath?: string;
 
     variables?: Array<ITemplateVariable>;